+++ src/analyser/Para.java
@@ -2,10 +2,9 @@ package analyser;
 
 import analyser.types.Type;
 
-import java.util.OptionalInt;
 import java.util.OptionalLong;
 
-public record Para(Type paraType, String paraName, OptionalLong offset, OptionalInt bitWidth) {
+public record Para(Type paraType, String paraName, OptionalLong offset, OptionalLong bitWidth) {
     @Override
     public String toString() {
         return "Para{" +
